chore: reinserted legend in topic info component, redefined topic name
