+++ src/lang/de.json
@@ -215,7 +215,7 @@
   "ch.sbb.zweitausbildung.tourist.pois.railaway-name": "Touristische Reiseziele mit RailAway",
   "ch.sbb.zweitausbildung.tourist.routes.group": "Touristische Strecken",
   "ch.sbb.zweitausbildung.tourist.routes.group-desc": "Die Touristischen Strecken umfassen alle Panorama-Strecken sowie weitere touristisch attraktive Zugstrecken.",
-  "ch.sbb.beleuchtungsstaerken": "Beleuchtungsstärken",
+  "ch.sbb.beleuchtungsstaerken": "Beleuchtung Bahnhofklassen",
   "ch.sbb.beleuchtungsstaerken1": "Bahnhofklasse 1",
   "ch.sbb.beleuchtungsstaerken2a": "Bahnhofklasse 2a",
   "ch.sbb.beleuchtungsstaerken2b": "Bahnhofklasse 2b",
