+++ src/lang/en.json
@@ -219,7 +219,7 @@
   "ch.sbb.zweitausbildung.tourist.pois.railaway-name": "Tourist destinations with RailAway",
   "ch.sbb.zweitausbildung.tourist.routes.group": "Tourist routes",
   "ch.sbb.zweitausbildung.tourist.routes.group-desc": "The tourist routes cover all panoramic routes as well as further attractive tourist train routes.",
-  "ch.sbb.beleuchtungsstaerken": "Lighting levels",
+  "ch.sbb.beleuchtungsstaerken": "Station classes for lighting",
   "ch.sbb.beleuchtungsstaerken1": "Station class 1",
   "ch.sbb.beleuchtungsstaerken2a": "Station class 2a",
   "ch.sbb.beleuchtungsstaerken2b": "Station class 2b",
