+++ src/lang/fr.json
@@ -235,7 +235,7 @@
   "ch.sbb.zweitausbildung.tourist.pois.railaway-name": "Destinations touristiques avec RailAway",
   "ch.sbb.zweitausbildung.tourist.routes.group": "Lignes touristiques",
   "ch.sbb.zweitausbildung.tourist.routes.group-desc": "On entend par «lignes touristiques» les lignes panoramiques et autres lignes touristiques.",
-  "ch.sbb.beleuchtungsstaerken": "Niveaux d'éclairement",
+  "ch.sbb.beleuchtungsstaerken": "Classes de gare pour l'éclairage",
   "ch.sbb.beleuchtungsstaerken1": "Classe de gare 1",
   "ch.sbb.beleuchtungsstaerken2a": "Classe de gare 2a",
   "ch.sbb.beleuchtungsstaerken2b": "Classe de gare 2b",
