+++ src/lang/it.json
@@ -236,7 +236,7 @@
   "ch.sbb.zweitausbildung.tourist.pois.railaway-name": "Mete turistiche con RailAway",
   "ch.sbb.zweitausbildung.tourist.routes.group": "Linee turistiche",
   "ch.sbb.zweitausbildung.tourist.routes.group-desc": "Le linee turistiche comprendono le linee panoramiche e altre linee turistiche.",
-  "ch.sbb.beleuchtungsstaerken": "Livelli di illuminamento",
+  "ch.sbb.beleuchtungsstaerken": "Classi stazioni per l'illuminazione",
   "ch.sbb.beleuchtungsstaerken1": "Classe stazione 1",
   "ch.sbb.beleuchtungsstaerken2a": "Classe stazione 2a",
   "ch.sbb.beleuchtungsstaerken2b": "Classe stazione 2b",
