+++ src/layerInfos/BeleuchtungTopicInfo/BeleuchtungTopicInfo.js
@@ -1,5 +1,6 @@
 import React from 'react';
 import { useTranslation } from 'react-i18next';
+import BeleuchtungLegende from './BeleuchtungLegende';
 
 const comps = {
   de: (
@@ -7,7 +8,8 @@ const comps = {
       Werkzeug für Grundlagengespräche u.a. mit Behörden, Gemeinden, Projekt-
       und Anlagenverantwortlichen über die Spezifikationen von
       Beleuchtungsanlagen.
-      <p>Bahnhofklasse gemäss VöV RTE 26201.</p>
+      <BeleuchtungLegende />
+      <p>Bahnhofklassen gemäss VöV RTE 26201.</p>
       <p>
         Verantwortlich: SBB Infrastruktur, Anlagenmanagement für Beleuchtung,
         &nbsp;
@@ -23,6 +25,7 @@ const comps = {
       Outil pour les discussions de base, entre autres avec les autorités, les
       communes, les responsables de projets et d&apos;installations, sur les
       spécifications des installations d&apos;éclairage.
+      <BeleuchtungLegende />
       <p>Classes de gare selon VÖV RTE 26201.</p>
       <p>
         Responsable: CFF Infrastructure, gestion des installations pour
@@ -39,6 +42,7 @@ const comps = {
       Strumento per discussioni di base sulle specifiche degli impianti di
       illuminazione con autorità, comuni, responsabili di progetti e
       installazioni, tra gli altri.
+      <BeleuchtungLegende />
       <p>Classi di stazioni secondo VÖV RTE 26201.</p>
       <p>
         Responsabile: SBB Infrastructure, gestione degli asset per
@@ -55,6 +59,7 @@ const comps = {
       Tool for basic discussions with, among others, authorities,
       municipalities, project and plant managers on the specifications of
       lighting installations.
+      <BeleuchtungLegende />
       <p>Station classes according to VÖV RTE 26201.</p>
       <p>
         Responsibility lies with: SBB Infrastructure, asset management for
